+++ lib/src/components/form/fields/select.dart
@@ -2,6 +2,7 @@ import 'dart:ui';
 
 import 'package:flutter/material.dart';
 import 'package:shadcn_ui/src/components/form/field.dart';
+import 'package:shadcn_ui/src/components/popover.dart';
 import 'package:shadcn_ui/src/components/select.dart';
 import 'package:shadcn_ui/src/raw_components/portal.dart';
 import 'package:shadcn_ui/src/theme/components/decorator.dart';
@@ -42,6 +43,13 @@ class ShadSelectFormField<T> extends ShadFormBuilderField<T> {
     ScrollController? scrollController,
     ShadAnchorBase? anchor,
     ImageFilter? filter,
+    ShadPopoverController? controller,
+
+    /// {@macro select.header}
+    Widget? header,
+
+    /// {@macro select.footer}
+    Widget? footer,
   }) : super(
           decorationBuilder: (context) =>
               (ShadTheme.of(context).selectTheme.decoration ??
@@ -71,6 +79,9 @@ class ShadSelectFormField<T> extends ShadFormBuilderField<T> {
               showScrollToBottomChevron: showScrollToBottomChevron,
               scrollController: scrollController,
               filter: filter,
+              controller: controller,
+              header: header,
+              footer: footer,
             );
           },
         );
@@ -116,6 +127,13 @@ class ShadSelectFormField<T> extends ShadFormBuilderField<T> {
     EdgeInsets? searchPadding,
     Widget? search,
     bool? clearSearchOnClose,
+    ShadPopoverController? controller,
+
+    /// {@macro select.header}
+    Widget? header,
+
+    /// {@macro select.footer}
+    Widget? footer,
   }) : super(
           decorationBuilder: (context) =>
               (ShadTheme.of(context).selectTheme.decoration ??
@@ -152,6 +170,9 @@ class ShadSelectFormField<T> extends ShadFormBuilderField<T> {
               searchPadding: searchPadding,
               search: search,
               clearSearchOnClose: clearSearchOnClose,
+              controller: controller,
+              header: header,
+              footer: footer,
             );
           },
         );
@@ -195,6 +216,13 @@ class ShadSelectFormField<T> extends ShadFormBuilderField<T> {
     EdgeInsets? searchPadding,
     Widget? search,
     bool? clearSearchOnClose,
+    ShadPopoverController? controller,
+
+    /// {@macro select.header}
+    Widget? header,
+
+    /// {@macro select.footer}
+    Widget? footer,
   })  : assert(
           variant == ShadSelectVariant.primary || onSearchChanged != null,
           'onSearchChanged must be provided when variant is search',
@@ -235,6 +263,9 @@ class ShadSelectFormField<T> extends ShadFormBuilderField<T> {
               searchPadding: searchPadding,
               search: search,
               clearSearchOnClose: clearSearchOnClose,
+              controller: controller,
+              header: header,
+              footer: footer,
             );
           },
         );